# apps/web/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminSkillListData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminSkillData"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AdminSkillListData",
        "type": "object"
      },
      "AdminSkillListResponse": {
        "properties": {
          "code": {
            "const": 0,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/AdminSkillListData"
          },
          "msg": {
            "const": "SUCCESS",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "code",
          "msg",
          "data"
        ],
        "title": "AdminSkillListResponse",
        "type": "object"
      },
...
      "AuthenticationErrorResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/UnauthorizedErrorResponse"
          },
          {
            "$ref": "#/components/schemas/SessionInvalidErrorResponse"
          }
        ],
        "description": "Bearer 缺失、无效或会话失效。",
        "title": "AuthenticationErrorResponse"
      },
...
      "Body_uploadSkill": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_uploadSkill",
        "type": "object"
      },
...
      "ConflictErrorResponse": {
        "properties": {
          "code": {
            "const": 10005,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "msg": {
            "const": "CONFLICT",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data",
          "code",
          "msg"
        ],
        "title": "ConflictErrorResponse",
        "type": "object"
      },
...
      "HTTPErrorResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/InternalErrorResponse"
          },
          {
            "$ref": "#/components/schemas/ValidationErrorResponse"
          },
          {
            "$ref": "#/components/schemas/UnauthorizedErrorResponse"
          },
          {
            "$ref": "#/components/schemas/ForbiddenErrorResponse"
          },
          {
            "$ref": "#/components/schemas/NotFoundErrorResponse"
          },
          {
            "$ref": "#/components/schemas/ConflictErrorResponse"
          },
          {
            "$ref": "#/components/schemas/RateLimitedErrorResponse"
          },
          {
            "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
          },
          {
            "$ref": "#/components/schemas/RegistrationDisabledErrorResponse"
          },
          {
            "$ref": "#/components/schemas/AccountIdentifierUnavailableErrorResponse"
          },
          {
            "$ref": "#/components/schemas/RegistrationRateLimitedErrorResponse"
          },
          {
            "$ref": "#/components/schemas/InvalidCredentialsErrorResponse"
          },
          {
            "$ref": "#/components/schemas/LoginRateLimitedErrorResponse"
          },
          {
            "$ref": "#/components/schemas/SessionInvalidErrorResponse"
          },
          {
            "$ref": "#/components/schemas/CurrentPasswordInvalidErrorResponse"
          },
          {
            "$ref": "#/components/schemas/MediaInvalidErrorResponse"
          },
          {
            "$ref": "#/components/schemas/MediaTooLargeErrorResponse"
          }
        ],
        "description": "所有稳定错误响应的联合类型，用于未预见状态的兜底契约。",
        "title": "HTTPErrorResponse"
      },
...
      "InternalErrorResponse": {
        "properties": {
          "code": {
            "const": -1,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "msg": {
            "const": "INTERNAL_ERROR",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data",
          "code",
          "msg"
        ],
        "title": "InternalErrorResponse",
        "type": "object"
      },
...
      "NotFoundErrorResponse": {
        "properties": {
          "code": {
            "const": 10004,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "msg": {
            "const": "NOT_FOUND",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data",
          "code",
          "msg"
        ],
        "title": "NotFoundErrorResponse",
        "type": "object"
      },
...
      "ServiceUnavailableErrorResponse": {
        "properties": {
          "code": {
            "const": 10007,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "msg": {
            "const": "SERVICE_UNAVAILABLE",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data",
          "code",
          "msg"
        ],
        "title": "ServiceUnavailableErrorResponse",
        "type": "object"
      },
...
      "SkillDeleteResponse": {
        "properties": {
          "code": {
            "const": 0,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/SkillDeleteData"
          },
          "msg": {
            "const": "SUCCESS",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "code",
          "msg",
          "data"
        ],
        "title": "SkillDeleteResponse",
        "type": "object"
      },
      "SkillEnabledRequest": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "SkillEnabledRequest",
        "type": "object"
      },
      "SkillListData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SkillData"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "SkillListData",
        "type": "object"
      },
      "SkillListResponse": {
        "properties": {
          "code": {
            "const": 0,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/SkillListData"
          },
          "msg": {
            "const": "SUCCESS",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "code",
          "msg",
          "data"
        ],
        "title": "SkillListResponse",
        "type": "object"
      },
...
      "SkillResourceResponse": {
        "properties": {
          "code": {
            "const": 0,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/SkillResourceData"
          },
          "msg": {
            "const": "SUCCESS",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "code",
          "msg",
          "data"
        ],
        "title": "SkillResourceResponse",
        "type": "object"
      },
      "SkillResponse": {
        "properties": {
          "code": {
            "const": 0,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/SkillData"
          },
          "msg": {
            "const": "SUCCESS",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "code",
          "msg",
          "data"
        ],
        "title": "SkillResponse",
        "type": "object"
      },
      "SkillUpdateRequest": {
        "properties": {
          "current_version_id": {
            "format": "uuid",
            "title": "Current Version Id",
            "type": "string"
          },
          "skill_md_text": {
            "maxLength": 1048576,
            "minLength": 1,
            "title": "Skill Md Text",
            "type": "string"
          }
        },
        "required": [
          "skill_md_text",
          "current_version_id"
        ],
        "title": "SkillUpdateRequest",
        "type": "object"
      },
...
      "UnauthorizedErrorResponse": {
        "properties": {
          "code": {
            "const": 10002,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "msg": {
            "const": "UNAUTHORIZED",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data",
          "code",
          "msg"
        ],
        "title": "UnauthorizedErrorResponse",
        "type": "object"
      },
...
      "ValidationErrorResponse": {
        "properties": {
          "code": {
            "const": 10001,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "msg": {
            "const": "VALIDATION_ERROR",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data",
          "code",
          "msg"
        ],
        "title": "ValidationErrorResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/v1/skills": {
      "get": {
        "operationId": "listAdminSkills",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSkillListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Admin Skills",
        "tags": [
          "admin-skills"
        ]
      }
    },
...
    "/api/v1/skills": {
      "get": {
        "operationId": "listSkills",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "客户端应使用 Bearer 认证方案",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Skills",
        "tags": [
          "skills"
        ]
      },
      "post": {
        "operationId": "uploadSkill",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadSkill"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "客户端应使用 Bearer 认证方案",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Post Skill",
        "tags": [
          "skills"
        ]
      }
    },
    "/api/v1/skills/{skill_id}": {
      "delete": {
        "operationId": "deleteSkill",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "客户端应使用 Bearer 认证方案",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Skill Endpoint",
        "tags": [
          "skills"
        ]
      },
      "get": {
        "operationId": "getSkill",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "客户端应使用 Bearer 认证方案",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Skill Detail",
        "tags": [
          "skills"
        ]
      }
    },
    "/api/v1/skills/{skill_id}/enabled": {
      "patch": {
        "operationId": "setSkillEnabled",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillEnabledRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "客户端应使用 Bearer 认证方案",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Skill Enabled",
        "tags": [
          "skills"
        ]
      }
    },
    "/api/v1/skills/{skill_id}/resource": {
      "get": {
        "operationId": "getSkillResource",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "maxLength": 1000,
              "minLength": 1,
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResourceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "客户端应使用 Bearer 认证方案",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Skill Resource",
        "tags": [
          "skills"
        ]
      }
    },
    "/api/v1/skills/{skill_id}/skill-md": {
      "put": {
        "operationId": "updateSkillMarkdown",
        "parameters": [
          {
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "客户端应使用 Bearer 认证方案",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Put Skill Markdown",
        "tags": [
          "skills"
        ]
      }
    },
...
  }
}